fix(agent): keep changed validation within limits
